core: DRM compatibility improvements
This is a fix for enable the DRM backend in westeros.

It ensures that the renderer does not exits when a valid depth
is reported by the EGL libraries, but tries to use the first
returned configuration.

For DRM we also need to call a DRM specific swap buffers, which
is implemented by extending the platform interface.

Signed-off-by: Zoltan Kuscsik <kuscsik@gmail.com>